+++ app/Exceptions/InvalidTodoTransition.php
@@ -2,6 +2,8 @@
 
 namespace App\Exceptions;
 
+use App\Enums\TodoStatus;
+use App\Enums\TodoTransition;
 use RuntimeException;
 
 /**
@@ -14,16 +16,49 @@ final class InvalidTodoTransition extends RuntimeException
 {
     public static function cannotCompleteArchived(): self
     {
-        return new self('Archived tasks must be unarchived before they can be completed.');
+        return new self((string) __('todos.exceptions.cannot_complete_archived'));
     }
 
     public static function cannotReopenArchived(): self
     {
-        return new self('Archived tasks must be unarchived before they can be reopened.');
+        return new self((string) __('todos.exceptions.cannot_reopen_archived'));
     }
 
     public static function cannotEditArchived(): self
     {
-        return new self('Archived tasks must be unarchived before they can be edited.');
+        return new self((string) __('todos.exceptions.cannot_edit_archived'));
+    }
+
+    public static function cannotCompleteTrashed(): self
+    {
+        return new self((string) __('todos.exceptions.cannot_complete_trashed'));
+    }
+
+    public static function cannotReopenTrashed(): self
+    {
+        return new self((string) __('todos.exceptions.cannot_reopen_trashed'));
+    }
+
+    public static function cannotArchiveTrashed(): self
+    {
+        return new self((string) __('todos.exceptions.cannot_archive_trashed'));
+    }
+
+    public static function cannotUnarchiveTrashed(): self
+    {
+        return new self((string) __('todos.exceptions.cannot_unarchive_trashed'));
+    }
+
+    public static function cannotEditTrashed(): self
+    {
+        return new self((string) __('todos.exceptions.cannot_edit_trashed'));
+    }
+
+    public static function invalid(TodoStatus $status, TodoTransition $transition): self
+    {
+        return new self((string) __('todos.exceptions.invalid_transition', [
+            'status' => $status->value,
+            'transition' => str_replace('_', ' ', $transition->value),
+        ]));
     }
 }